requirements.txt
numpy>=1.23.0
Pillow>=9.0.0
torch>=2.0.0
torchvision>=0.15.0
torchsummary>=1.5.1
scikit-image>=0.19.0
opencv-python>=4.7.0
tqdm>=4.65.0
pyyaml>=6.0
scikit-learn>=1.2.0
matplotlib>=3.6.0
albumentations>=1.3.0
psutil>=5.9.0